- Remove dependency from commons lang 3. Instead of using commons lang 3, we will use commons lang 2.6 and we will mark it as dependency in pom file
- Remove direct dependency of hudson remoting module from our pom. I don't know why we were marking removing as direct dependency in pom file. Instead of doing that IMHO it is better to use the remoting module provided with Jenkins out of the box.

- Remove transitive dependencies of :
	- jackon-annotations
	- commons-codec
	- commons-logging
	- bouncycastle
	- joda-time
	- google guava.

It was impossible for me to find out why we have this jars packaged in our HPI file. When I removed them the code compiles ok and in runtime I haven't found where they could be needed. With no other references  and information I think that it is better to remove them in order to know what third party libraries we are really needing.